move transformation tests in their own packages
